% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastTS_cv.R
\name{fastTS_cv}
\alias{fastTS_cv}
\alias{plot.fastTS_cv}
\alias{coef.fastTS_cv}
\alias{print.fastTS_cv}
\alias{summary.fastTS_cv}
\title{Use k-fold CV with fastTS}
\usage{
fastTS_cv(
  y,
  X = NULL,
  n_lags_max,
  gamma = c(0, 2^(-2:4)),
  ptrain = 1,
  cv_folds = 10,
  pf_eps = 0.01,
  w_endo,
  w_exo,
  ncvreg_args = list(penalty = "lasso", returnX = FALSE, lambda.min = 0.001, nlambda =
    100)
)

\method{plot}{fastTS_cv}(x, log.l = TRUE, ...)

\method{coef}{fastTS_cv}(object, ...)

\method{print}{fastTS_cv}(x, ...)

\method{summary}{fastTS_cv}(object, ...)
}
\arguments{
\item{y}{univariate time series outcome}

\item{X}{matrix of predictors (no intercept)}

\item{n_lags_max}{maximum number of lags to consider}

\item{gamma}{vector of exponent for weights}

\item{ptrain}{prop. to leave out for test data}

\item{cv_folds}{number of cross-validation folds (defaults to 10)}

\item{pf_eps}{penalty factors below this will be set to zero}

\item{w_endo}{optional pre-specified weights for endogenous terms}

\item{w_exo}{optional pre-specified weights for exogenous terms (see details)}

\item{ncvreg_args}{additional args to pass through to ncvreg}

\item{x}{a fastTS_cv object}

\item{log.l}{Should the x-axis (lambda) be logged?}

\item{...}{passed to downstream functions}

\item{object}{a fastTS_cv object}

\item{choose}{which criterion to use for lambda selection (AICc, BIC, or all)}
}
\value{
A list of class \code{fastTS_cv} with elements

  \item{fit}{a fit at the best CV estimate of gamma}
  \item{ncvreg_args}{arguments passed to ncvreg}
  \item{gamma}{the (negative) exponent on the penalty weights, one
  for each fit}
  \item{n_lags_max}{the maximum number of lags}
  \item{y}{the time series}
  \item{X}{the utilized matrix of exogenous features}
  \item{oos_results}{results on test data using best of fits}
  \item{train_idx}{index of observations used in training data}
  \item{cve_mean}{mean CV error for each gamma}
  \item{best_gamma}{best gamma}
  \item{best_lambda}{best lambda}
  \item{best_pf}{best penalty factors}

x invisibly

a vector of model coefficients

x (invisibly)

the summary object produced by ncvreg
  evaluated at the best tuning parameter combination
  (best AICc).
}
\description{
Use k-fold CV with fastTS
}
\details{
This function performs similarly to \code{fastTS}, except is
  designed for use with cross-validation. The additional need to
  cross-validate makes the PACF function, used in determining the weights
  more difficult as this must be wrapped into the CV procedure, and the PACF
  needs to be recomputed each time. Therefore this function takes potentially
  considerably longer than \code{fastTS}.

  The benefit of this function is that the user is not dependent on AICc or BIC for
  model tuning, and obtains extra-sample estimates of the model error which
  may be useful for forecasting.

  This function is currently experimental. It is not recommended for use at this time.
}
\examples{
data("LakeHuron")
fit_LH <- fastTS_cv(LakeHuron)
fit_LH
coef(fit_LH)
plot(fit_LH)

}
\seealso{
predict.fastTS_cv
}
