% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{penalty_scaler}
\alias{penalty_scaler}
\title{Penalty Scaling Function for parametric penalty weights}
\usage{
penalty_scaler(lag, m, r, plot = TRUE, log = TRUE)
}
\arguments{
\item{lag}{a vector of lags for which to calculate the penalty function}

\item{m}{a vector of seasonality modes}

\item{r}{a vector of dim (m + 1) for the factor penalties on c(m, time)}

\item{plot}{logical; whether to plot the penalty function}

\item{log}{logical; whether to return the log of the penalty function}
}
\description{
Penalty Scaling Function for parametric penalty weights
}
