% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict.srlTS}
\alias{predict.srlTS}
\title{Predict function for srlTS object}
\usage{
\method{predict}{srlTS}(object, n_ahead = 1, X_test, y_test, cumulative = 0, ...)
}
\arguments{
\item{object}{an srlTS object}

\item{n_ahead}{number of times ahead to predict by iteration}

\item{X_test}{a matrix exogenous features}

\item{y_test}{the test series, for future predictions if n_ahead <
nrow(X_test) (optional)}

\item{cumulative}{should cumulative (rolling) sums be returned (integer
indicating number of times to sum)}

\item{...}{currently unused}
}
\description{
Predict function for srlTS object
}
