% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseR_ts.R
\name{sparseR_ts}
\alias{sparseR_ts}
\title{Perform time series ranked sparsity methods}
\usage{
sparseR_ts(
  y,
  X = NULL,
  n_lags_max,
  gamma,
  ptrain = 0.8,
  pf_eps = 0.01,
  w_endo,
  w_exo,
  ncv_args = list(penalty = "lasso", returnX = FALSE, lambda.min = 0.001)
)
}
\arguments{
\item{y}{univariate time series outcome}

\item{X}{matrix of predictors (no intercept)}

\item{n_lags_max}{maximum number of lags to consider}

\item{gamma}{vector of exponent for weights}

\item{ptrain}{prop. to leave out for test data}

\item{pf_eps}{penalty factors below this will be set to zero}

\item{w_endo}{optional pre-specified weights for endogenous terms}

\item{w_exo}{optional pre-specified weights for exogenous terms}

\item{ncv_args}{additional args to pass through to ncvreg}
}
\description{
Perform time series ranked sparsity methods
}
