% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srlTS.R
\name{srlTS}
\alias{srlTS}
\alias{plot.srlTS}
\alias{coef.srlTS}
\alias{predict.srlTS}
\alias{print.srlTS}
\title{Perform time series ranked sparsity methods}
\usage{
srlTS(
  y,
  X = NULL,
  n_lags_max,
  gamma,
  ptrain = 0.8,
  pf_eps = 0.01,
  w_endo,
  w_exo,
  ncvreg_args = list(penalty = "lasso", returnX = FALSE, lambda.min = 0.001)
)
}
\arguments{
\item{y}{univariate time series outcome}

\item{X}{matrix of predictors (no intercept)}

\item{n_lags_max}{maximum number of lags to consider}

\item{gamma}{vector of exponent for weights}

\item{ptrain}{prop. to leave out for test data}

\item{pf_eps}{penalty factors below this will be set to zero}

\item{w_endo}{optional pre-specified weights for endogenous terms}

\item{w_exo}{optional pre-specified weights for exogenous terms}

\item{ncvreg_args}{additional args to pass through to ncvreg}
}
\description{
Perform time series ranked sparsity methods
}
